+++ module-market/src/main/java/com/market/board/repository/MarketBoardContentRepository.java
@@ -3,5 +3,5 @@ package com.market.board.repository;
 import com.market.board.entity.MarketBoardContent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MarketBoardContentRepository extends JpaRepository<MarketBoardContent, Integer> {
+public interface MarketBoardContentRepository extends JpaRepository<MarketBoardContent, Long> {
 }
